refactor CheckRemainingTime feature

+++ Features/Attempts/CheckRemainingTime/Commands/UpdateAttemptToTimedOutCommand.cs
@@ -3,5 +3,5 @@ using MediatR;
 
 namespace exam_system.Features.Attempts.CheckRemainingTime.Commands
 {
-    public record UpdateAttemptToTimedOutCommand(Guid AttemptId, decimal Score, bool Passed): IRequest<RequestResponse<Unit>>;
+    public record UpdateAttemptToTimedOutCommand(Guid AttemptId, double Score, bool Passed): IRequest<RequestResponse<Unit>>;
 }
